Make block device naming on CentOS consistant with Ubuntu.
There was a bug introduced in the RHEL/CentOS 6.1 release
that changes how block devices are named on Xen. This setting
forces 6.0 behavior which is consistant with Ubuntu and other
distros running a 3.x series Linux kernel. The bug is discussed
at https://bugzilla.redhat.com/show_bug.cgi?id=729586.

Signed-off-by: Tony Hansmann <thansmann@pivotallabs.com>